requirements.txt
openai>=1.16.0
python-dotenv>=0.21.0
